+++ transform/lib/types.js
@@ -1,3 +1,5 @@
+import * as asc from "assemblyscript/dist/assemblyscript.js";
+export const NodeKind = asc.NodeKind;
 export var PropertyFlags;
 (function (PropertyFlags) {
     PropertyFlags[PropertyFlags["OmitNull"] = 0] = "OmitNull";